Update Category.java
Make 'Id' to Long type

+++ src/main/java/ac/kr/jejunu/common/model/Category.java
@@ -16,6 +16,6 @@ import javax.persistence.Id;
 public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer category_no; //식별자
+    private Long category_no; //식별자
     private String category_name; //카테고리 이름
 }
